Fix counting extra lines when lines wrap
Previously, extra lines were counted only by the number of line breaks
in them, so the number of lines cleared would be incorrect if any lines
wrapped. Additionally, prevLineCount could be wrong if the size of the
terminal changes between renders.

To fix this, the number of lines to clear is calculated immediately
before clearing them, and the 'string-width' library is used to
determine accurate widths. This is the same library used by 'wrap-ansi',
so it should be reliable, and it is not actually a new dependency.